requirements.txt
Flask
langchain
langchain-community
together
pandas
chromadb
python-dotenv